Fix vga printing - Colour needs to be returned as a uint16_t to account for the colour and background

+++ kernel/arch/i386/vga.h
@@ -25,6 +25,6 @@ static inline uint8_t vga_entry_colour(enum vga_colour fg, enum vga_colour bg) {
     return fg | (bg << 4);                                                       // NOLINT(hicpp-signed-bitwise,cppcoreguidlines-avoid-magic-numbers)
 }
 
-static inline uint8_t vga_entry(unsigned char character, enum vga_colour colour) { // NOLINT(modernize-use-trailing-return-type,clang-diagnostic-unused-function)
+static inline uint16_t vga_entry(unsigned char character, enum vga_colour colour) { // NOLINT(modernize-use-trailing-return-type,clang-diagnostic-unused-function)
     return static_cast<uint16_t>(character) | static_cast<uint16_t>(colour << 8);  // NOLINT(hicpp-signed-bitwise,cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
 }
